day8 part2

+++ src/main/java/Main.java
@@ -11,7 +11,7 @@ public class Main {
     public static void main(String[] args) {
         String file = "C:\\Users\\Lenovo\\Desktop\\study\\AdventOfCode2023\\src\\main\\resources\\day8\\input.txt";
         try {
-            int result = CamelMap.totalSteps(file);
+            long result = CamelMap.totalSteps(file);
             System.out.println(result);
         } catch (IOException e) {
